Refactoring: move each reducer to separate file
