+++ src/store/reducers/root-reducer.js
@@ -1,12 +1,10 @@
 import { combineReducers } from 'redux';
 
-import {
-  sortByFilter,
-  movies,
-  isLoaded,
-  searchByFilter,
-  searchMoviesValue
-} from './reducers';
+import { isLoaded } from './is-loaded';
+import { movies } from './movies';
+import { searchByFilter } from './search-by-filter';
+import { searchMoviesValue } from './search-movies-value';
+import { sortByFilter } from './sort-by-filter';
 
 export const rootReducer = combineReducers({
   sortByFilter,
